[#152975373] Como consumidor, eu posso obter promoções próximas a minha localidade, para maior comodidade.

+++ app/src/main/java/projetaobcc20172/com/projetopetemfoco/activity/FiltroEstabelecimentoActivity.java
@@ -18,7 +18,7 @@ public class FiltroEstabelecimentoActivity extends Activity implements View.OnCl
 
     private CheckBox cbProx;
     private CheckBox cbAvaliacao;
-    private byte raio;
+    private int raio;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
